formulae_1_batch_pipepline/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Data processing
pandas==2.1.4
numpy==1.26.2

# API requests
requests>=2.32.0

# PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Environment variables
python-dotenv>=0.21.0